% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mcp_tools}
\alias{mcp_tools}
\title{MCP tools for the RStudio editor}
\usage{
mcp_tools()
}
\value{
A list of \code{ellmer::ToolDef} objects suitable for
\code{\link[mcptools:mcp_server]{mcptools::mcp_server()}}.
}
\description{
MCP tools for the RStudio editor
}
